refactor: edit the order of the properties
