+++ src/components/About/About.css
@@ -16,7 +16,7 @@
 }
 
 .about__title-description {
-  font-family: Roboto Slab, monospace;
+  font-family: Roboto Slab, Georgia, 'Times New Roman', Times, monospace;
   font-weight: normal;
   font-size: 40px;
   margin-top: 52px;
@@ -24,6 +24,8 @@
 }
 
 .about__text-description {
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
+  font-weight: 400;
   font-size: 18px;
   line-height: 24px;
   margin-top: 0;
