+++ src/components/Button/Button.css
@@ -34,6 +34,7 @@
 }
 
 .button__text {
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
   overflow: hidden;
   text-overflow: ellipsis;
   white-space: nowrap;
