+++ src/components/ButtonBurgerMenu/ButtonBurgerMenu.js
@@ -2,15 +2,17 @@ import React from 'react';
 import './ButtonBurgerMenu.css';
 
 function ButtonBurgerMenu(props) {
-  const buttonBurgerMenu = `button-burger-menu ${(props.pathname === '/' || props.isMenuMobile)
+  const buttonBurgerMenu = `button-burger-menu__button ${(props.pathname === '/' || props.isMenuMobile)
     ? ''
     : 'button-burger-menu_saved-news'}`;
 
   return (
-    <button
-      className={`${buttonBurgerMenu}  ${props.classBurgerMenu}`}
-      onClick={props.showMenu}
-    />
+    <div className="button-burger-menu">
+      <button
+        className={`${buttonBurgerMenu}  ${props.classBurgerMenu}`}
+        onClick={props.showMenu} type='button'
+      />
+    </div>
   );
 }
