+++ src/components/Footer/Footer.css
@@ -32,12 +32,15 @@
 }
 
 .footer__link {
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
   display: flex;
   color: #1a1b22;
   font-size: 18px;
 }
 
 .footer__copyright {
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
+  font-weight: 400;
   display: flex;
   color: #B6BCBF;
   font-size: 16px;
