+++ src/components/Link/Link.css
@@ -1,3 +1,4 @@
 .link {
   text-decoration: none;
+  cursor: pointer;
 }
