+++ src/components/Navigation/Navigation.css
@@ -12,6 +12,7 @@
 }
 
 .nav__text {
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
   font-style: normal;
   font-weight: 500;
   font-size: 18px;
@@ -39,7 +40,6 @@
   }
 
   .nav__text {
-    font-style: normal;
     font-size: 16px;
     line-height: 24px;
     margin-right: 18px;
