+++ src/components/NewsCard/NewsCard.css
@@ -28,6 +28,7 @@
 }
 
 .news-card__items-in-img {
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
   padding: 0 12px;
   font-style: normal;
   font-weight: 500;
@@ -50,14 +51,8 @@
   cursor: initial;
 }
 
-@media screen and (max-width: 960px) {
-  .news-card__keyword-container {
-    top: 8px;
-    left: 8px;
-  }
-}
-
 .news-card__keyword {
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
   display: block;
   font-size: 14px;
   overflow: hidden;
@@ -74,22 +69,6 @@
   justify-content: flex-end;
   z-index: 2;
   margin: auto;
-}
-
-@media screen and (max-width: 960px) {
-  .news-card__image-container {
-    top: 8px;
-    right: 8px;
-  }
-}
-
-.news-card__mark {
-  position: absolute;
-  padding: 0;
-  width: 40px;
-  justify-content: center;
-  background-repeat: no-repeat;
-  background-position: center;
 }
 
 .news-card__mark_type_trashcan {
@@ -121,6 +100,87 @@
   right: 40px;
 }
 
+.news-card__mark {
+  position: absolute;
+  padding: 0;
+  width: 40px;
+  justify-content: center;
+  background-repeat: no-repeat;
+  background-position: center;
+}
+
+.news-card__mark:hover + .news-card__notice {
+  display: flex;
+}
+
+.news-card__text-container {
+  display: grid;
+  grid-template-rows: 55px auto 1fr 46px;
+  height: 100%;
+  padding-left: 24px;
+  padding-right: 24px
+}
+
+.news-card__date {
+  font-family: 'Source Sans Pro', Arial, Helvetica, sans-serif;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 18px;
+  line-height: 24px;
+  color: #B6BCBF;
+  padding: 18px 0 12px;
+  margin: 0;
+}
+
+.news-card__link {
+  color: #1A1B22;
+  text-decoration: none;
+}
+
+.news-card__title {
+  font-family: Roboto Slab, Georgia, 'Times New Roman', Times, monospace;
+  font-style: normal;
+  font-weight: normal;
+  overflow: hidden;
+  font-size: 26px;
+  line-height: 30px;
+  padding-bottom: 15px;
+  display: -webkit-box;
+  -webkit-box-orient: vertical;
+  -webkit-line-clamp: 2;
+  height: 40px;
+  margin: 0;
+}
+
+.news-card__source {
+  font-family: Roboto Slab, Georgia, 'Times New Roman', Times, monospace;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 16px;
+  line-height: 20px;
+  letter-spacing: 0.4px;
+  text-transform: uppercase;
+  color: #B6BCBF;
+  text-align: left;
+  padding: 0;
+  margin: 0;
+}
+
+.news-card__content {
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 16px;
+  line-height: 22px;
+  padding-bottom: 22px;
+  display: -webkit-box;
+  -webkit-box-orient: vertical;
+  -webkit-line-clamp: 5;
+  overflow: hidden;
+  height: 86px;
+  margin: 20px 0 21px;
+}
+
 
 @media screen and (max-width: 1427px) {
   .news-card__notice {
@@ -141,6 +201,48 @@
   }
 }
 
+
+@media screen and (max-width: 960px) {
+  .news-card__text-container {
+    grid-template-rows: 55px auto 1fr 44px;
+    padding-left: 16px;
+    padding-right: 16px;
+  }
+
+
+  .news-card__title {
+    font-size: 21px;
+    line-height: 24px;
+    padding-bottom: 15px;
+    -webkit-line-clamp: 3;
+    height: 60px;
+    margin: 0;
+  }
+
+  .news-card__content {
+    -webkit-line-clamp: 4;
+    padding-top: 2px;
+    height: 65px;
+    margin: 2px 0;
+  }
+
+  .news-card__source {
+    margin-top: 5px;
+    padding-top: 4px;
+  }
+
+  .news-card__image-container {
+    top: 8px;
+    right: 8px;
+  }
+
+  .news-card__keyword-container {
+    top: 8px;
+    left: 8px;
+  }
+}
+
+
 @media screen and (max-width: 710px) {
   .news-card__notice {
     right: 40px;
@@ -151,7 +253,13 @@
     right: 40px;
     top: 0;
   }
+
+  .news-card__title {
+    -webkit-line-clamp: 2;
+    height: 36px;
+  }
 }
+
 
 @media screen and (max-width: 500px) {
   .news-card__notice {
@@ -159,6 +267,7 @@
     top: 0;
   }
 }
+
 
 @media screen and (max-width: 395px) {
   .news-card__notice {
@@ -171,118 +280,3 @@
     top: 0;
   }
 }
-
-.news-card__mark:hover + .news-card__notice {
-  display: flex;
-}
-
-.news-card__text-container {
-  display: grid;
-  grid-template-rows: 55px auto 1fr 46px;
-  height: 100%;
-  padding-left: 24px;
-  padding-right: 24px
-}
-
-@media screen and (max-width: 960px) {
-  .news-card__text-container {
-    grid-template-rows: 55px auto 1fr 44px;
-    padding-left: 16px;
-    padding-right: 16px;
-  }
-}
-
-
-.news-card__date {
-  font-family: 'Source Sans Pro', sans-serif;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 24px;
-  color: #B6BCBF;
-  padding: 18px 0 12px;
-  margin: 0;
-}
-
-.news-card__link {
-  color: #1A1B22;
-  text-decoration: none;
-}
-
-.news-card__title {
-  font-family: Roboto Slab, monospace;
-  font-style: normal;
-  font-weight: normal;
-  overflow: hidden;
-  font-size: 26px;
-  line-height: 30px;
-  padding-bottom: 15px;
-  display: -webkit-box;
-  -webkit-box-orient: vertical;
-  -webkit-line-clamp: 2;
-  height: 40px;
-  margin: 0;
-}
-
-@media screen and (max-width: 960px) {
-  .news-card__title {
-    font-size: 21px;
-    line-height: 24px;
-    padding-bottom: 15px;
-    -webkit-line-clamp: 3;
-    height: 60px;
-    margin: 0;
-  }
-}
-
-@media screen and (max-width: 710px) {
-  .news-card__title {
-    -webkit-line-clamp: 2;
-    height: 36px;
-  }
-}
-
-.news-card__content {
-  font-style: normal;
-  font-weight: normal;
-  font-size: 16px;
-  line-height: 22px;
-  padding-bottom: 22px;
-  display: -webkit-box;
-  -webkit-box-orient: vertical;
-  -webkit-line-clamp: 5;
-  overflow: hidden;
-  height: 86px;
-  margin: 20px 0 21px;
-}
-
-@media screen and (max-width: 960px) {
-  .news-card__content {
-    -webkit-line-clamp: 4;
-    padding-top: 2px;
-    height: 65px;
-    margin: 2px 0;
-  }
-}
-
-.news-card__source {
-  font-family: Roboto Slab, monospace;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 16px;
-  line-height: 20px;
-  letter-spacing: 0.4px;
-  text-transform: uppercase;
-  color: #B6BCBF;
-  text-align: left;
-  padding: 0;
-  margin: 0;
-}
-
-@media screen and (max-width: 960px) {
-  .news-card__source {
-    margin-top: 5px;
-    padding-top: 4px;
-  }
-}
-
