+++ src/components/NewsCard/NewsCard.js
@@ -23,7 +23,8 @@ function NewsCard(props) {
       return (
         <>
           <button className="news-card__items-in-img news-card__mark news-card__mark_type_save"/>
-          <p className="news-card__items-in-img news-card__notice news-card__notice_main">Войдите, чтобы сохранять статьи</p>
+          <p className="news-card__items-in-img news-card__notice news-card__notice_main">Войдите, чтобы сохранять
+            статьи</p>
         </>
       );
     }
@@ -39,7 +40,8 @@ function NewsCard(props) {
 
   return (
     <li className="news-card">
-      <a title="Открыть статью в новой вкладке" href={props.url} className="link news-card__link-img" target="_blank" rel='noopener noreferrer'>
+      <a title="Открыть статью в новой вкладке" href={props.url} className="link news-card__link-img" target="_blank"
+         rel='noopener noreferrer'>
         <img className="news-card__image"
              src={props.urlToImage}
              alt={props.title}
@@ -60,7 +62,8 @@ function NewsCard(props) {
 
       <div className="news-card__text-container">
         <span className="news-card__date">{fullDate}</span>
-        <a title="Открыть статью в новой вкладке" className="news-card__link" href={props.url} target="_blank" rel='noopener noreferrer'>
+        <a title="Открыть статью в новой вкладке" className="news-card__link" href={props.url} target="_blank"
+           rel='noopener noreferrer'>
           <h3 className="news-card__title">{props.title}</h3>
         </a>
         <p className="news-card__content">{props.content}</p>
