+++ src/components/NewsCardList/NewsCardList.css
@@ -1,105 +1,11 @@
 .news-card-list {
-  background-color: #F5F6F7;
-}
-
-.news-card-list {
   max-width: calc(1440px + 208px);
   margin: 0 -104px;
-}
-
-@media screen and (max-width: 960px) {
-  .news-card-list {
-    margin: 0 -40px;
-  }
-}
-
-@media screen and (max-width: 500px) {
-  .news-card-list {
-    margin: 0 -16px;
-  }
+  background-color: #F5F6F7;
 }
 
 .news-card-list__container {
   margin: 0 104px;
-}
-
-
-@media screen and (max-width: 960px) {
-  .news-card-list__container {
-    margin: 0 40px;
-  }
-}
-
-@media screen and (max-width: 500px) {
-  .news-card-list__container {
-    margin: 0 16px;
-  }
-}
-
-.news-card-list__list {
-  list-style: none;
-  padding: 62px 0;
-  display: grid;
-  margin: 0 auto;
-  grid-template-columns: repeat(3, 1fr);
-  row-gap: 16px;
-  grid-column-gap: 16px;
-  -webkit-column-gap: 16px;
-  column-gap: 16px;
-  justify-content: center;
-  width: 100%;
-}
-
-.news-card-list__title {
-  font-family: Roboto Slab, monospace;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 40px;
-  line-height: 46px;
-  margin: 0;
-  padding: 80px 0 0;
-}
-
-.news-card-list__title_hidden {
-  display: none;
-}
-
-@media screen and (max-width: 960px) {
-  .news-card-list__list {
-    padding: 32px 0;
-    row-gap: 10px;
-    grid-column-gap: 10px;
-    -webkit-column-gap: 10px;
-    column-gap: 10px;
-  }
-
-  .news-card-list__title {
-    font-size: 30px;
-    line-height: 34px;
-    padding: 40px 0 0;
-  }
-}
-
-@media screen and (max-width: 710px) {
-  .news-card-list__list {
-    grid-template-columns: 400px;
-  }
-
-  .news-card-list__title {
-    width: 400px;
-    margin-left: auto;
-    margin-right: auto;
-  }
-}
-
-@media screen and (max-width: 432px) {
-  .news-card-list__list {
-    grid-template-columns: 1fr;
-  }
-
-  .news-card-list__title {
-    width: auto;
-  }
 }
 
 .news-card-list__button-position {
@@ -117,12 +23,60 @@
 .news-card-list__button-text {
   color: #1A1B22;
   font-style: normal;
-  font-weight: 500;
-  font-size: 18px;
   line-height: 24px;
 }
 
+.news-card-list__list {
+  list-style: none;
+  padding: 62px 0;
+  display: grid;
+  margin: 0 auto;
+  grid-template-columns: repeat(3, 1fr);
+  row-gap: 16px;
+  grid-column-gap: 16px;
+  -webkit-column-gap: 16px;
+  column-gap: 16px;
+  justify-content: center;
+  width: 100%;
+}
+
+.news-card-list__title {
+  font-family: Roboto Slab, Georgia, 'Times New Roman', Times, monospace;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 40px;
+  line-height: 46px;
+  margin: 0;
+  padding: 80px 0 0;
+}
+
+.news-card-list__title_hidden {
+  display: none;
+}
+
 @media screen and (max-width: 960px) {
+  .news-card-list {
+    margin: 0 -40px;
+  }
+
+  .news-card-list__list {
+    padding: 32px 0;
+    row-gap: 10px;
+    grid-column-gap: 10px;
+    -webkit-column-gap: 10px;
+    column-gap: 10px;
+  }
+
+  .news-card-list__title {
+    font-size: 30px;
+    line-height: 34px;
+    padding: 40px 0 0;
+  }
+
+  .news-card-list__container {
+    margin: 0 40px;
+  }
+
   .news-card-list__button-position {
     padding-bottom: 40px;
   }
@@ -133,7 +87,22 @@
   }
 }
 
+@media screen and (max-width: 710px) {
+  .news-card-list__list {
+    grid-template-columns: 400px;
+  }
+
+  .news-card-list__title {
+    width: 400px;
+    margin-left: auto;
+    margin-right: auto;
+  }
+}
+
 @media screen and (max-width: 500px) {
+  .news-card-list__container {
+    margin: 0 16px;
+  }
 
   .news-card-list__button {
     height: 56px;
@@ -142,5 +111,19 @@
 
   .news-card-list__list {
     padding-bottom: 24px;
+  }
+
+  .news-card-list {
+    margin: 0 -16px;
+  }
+}
+
+@media screen and (max-width: 432px) {
+  .news-card-list__list {
+    grid-template-columns: 1fr;
+  }
+
+  .news-card-list__title {
+    width: auto;
   }
 }
