+++ src/components/NotFound/NotFound.css
@@ -14,6 +14,8 @@
 }
 
 .not-found__text {
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
+  font-weight: 400;
   font-size: 18px;
   line-height: 24px;
   text-align: center;
