+++ src/components/Popups/PopupForm/PopupForm.css
@@ -15,10 +15,6 @@
   z-index: -1;
 }
 
-.popup__container {
-
-}
-
 .popup_opened {
   z-index: 10;
   pointer-events: auto;
@@ -72,7 +68,7 @@
 }
 
 .popup__title {
-  font-family: Roboto, Arial, Helvetica, sans-serif;
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
   font-size: 24px;
   line-height: 32px;
   font-weight: 900;
@@ -107,7 +103,11 @@
 }
 
 .popup__link:hover {
-  color: #347eff;
+  color: #347EFF;
+}
+
+.popup__link:focus {
+  color: #2A65CC;
 }
 
 .popup__button {
@@ -118,18 +118,22 @@
   margin: 42px 0 15px;
 }
 
+.popup__button:hover {
+  background-color: #347EFF;
+}
+
+.popup__button:focus {
+  background-color: #2A65CC;
+}
+
 .popup__button:disabled {
   color: #b6bcbf;
-  background-color: #e6e8eb;
+  background-color: #E6E8EB;
   pointer-events: none;
 }
 
-.popup__button:hover {
-  background-color: #e8e8e9;
-}
-
 .popup__button:active {
-  background-color: #2a65cc;
+  background-color: #2A65CC;
 }
 
 .popup__warning {
