+++ src/components/Popups/PopupInput/PopupInput.css
@@ -3,7 +3,7 @@
   background-color: white;
   border-radius: 100px;
   box-sizing: border-box;
-  font-family: Roboto, Arial, Helvetica, sans-serif;
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
   font-size: 12px;
   font-weight: 400;
   color: #2f71e5;
