+++ src/components/Popups/Register/Register.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PopupInput from '../PopupInput/PopupInput';
 import PopupForm from '../PopupForm/PopupForm';
-import useFormWithValidation from '../../../utils/validation';
+import useFormWithValidation from '../../../utils/useFormWithValidation';
 
 export default function Register(props) {
   const {
