+++ src/components/Preloader/Preloader.css
@@ -29,6 +29,7 @@
 }
 
 .preloader__title {
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
   font-size: 18px;
   line-height: 24px;
   text-align: center;
