+++ src/components/SearchForm/SearchForm.css
@@ -1,4 +1,3 @@
-/*Одновременно менять index.css  и .header__border*/
 .search-form {
   background: linear-gradient(rgba(0, 0, 0, .3), rgba(0, 0, 0, .3)), url("../../images/background.jpg") center;
   background-size: cover;
@@ -15,7 +14,7 @@
 }
 
 .search-form__title {
-  font-family: Roboto Slab, monospace;
+  font-family: Roboto Slab, Georgia, 'Times New Roman', Times, monospace;
   font-weight: normal;
   font-size: 60px;
   line-height: 64px;
@@ -23,6 +22,8 @@
 }
 
 .search-form__subtitle {
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
+  font-weight: 400;
   font-size: 18px;
   line-height: 24px;
   margin: 0;
@@ -36,16 +37,22 @@
   box-sizing: border-box;
   border-radius: 100px;
   background-color: white;
-  /*border: 1px solid #d1d2d6;*/
   margin-top: 68px;
 }
 
 .search-form__input {
+  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
+  font-weight: 400;
   border: none;
   box-sizing: border-box;
   width: 100%;
   border-radius: 100px;
   padding-left: 16px;
+  color: #1A1B22;
+}
+
+.search-form__input::placeholder {
+  color: #B6BCBF;
 }
 
 .search-form__button {
@@ -56,8 +63,14 @@
   align-self: flex-start;
 }
 
+.search-form__button:hover {
+  background-color: #347EFF;
+}
 
-/*Одновременно менять index.css  и .header__border*/
+.search-form__button:focus {
+  background-color: #2A65CC;
+}
+
 @media screen and (max-width: calc(960px)) {
   .search-form {
     margin-left: -40px;
@@ -99,14 +112,9 @@
 }
 
 @media screen and (max-width: 710px) {
-  /*Потому что position: fixed; */
   .search-form {
     margin-top: 0;
     height: 536px;
-  }
-
-  .search-form__container {
-
   }
 
   .search-form__container {
@@ -134,6 +142,5 @@
     margin-left: -16px;
     margin-right: -16px;
   }
-
 }
