+++ src/index.css
@@ -22,8 +22,6 @@ code {
 }
 
 .page {
-  font-family: 'Roboto', 'Inter', 'Arial', sans-serif;
-  font-weight: 400;
   max-width: 1440px;
   min-width: calc(320px - 32px);
   margin: 0 auto;
