--- src/utils/color.css
@@ -1,27 +1,0 @@
-.BLACK {
-  color: black;
-}
-
-.DARK-GRAY {
-  color: #1A1B22;
-}
-
-.LIGHT-GREY {
-  color: #B6BCBF;
-}
-
-.LIGHT-GREY-BGR {
-  color: #E6E8EB;;
-}
-
-.BLUE-PRESS {
-  color: #2A65CC;
-}
-
-.BLUE {
-  color: #2F71E5;
-}
-
-.BLUE-HOVER {
-  color: #347EFF;
-}
